Refactor CLI.py and utils.py files
